requirements.txt
aniso8601==7.0.0
attrs==21.2.0
brotli==1.0.9
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.6; python_version >= '3'
click==8.0.1
configargparse==1.5.3
fastapi==0.68.1
flask-basicauth==0.2.0
flask-cors==3.0.10
flask==2.0.1
gevent==21.8.0
geventhttpclient==1.5.2
google-api-core==2.1.0
google-auth==2.3.0
grpcio-tools~=1.30
googleapis-common-protos==1.53.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.2; platform_python_implementation == 'CPython'
grpc-google-iam-v1==0.12.3
grpcio==1.41.0
idna==3.2; python_version >= '3'
iniconfig==1.1.1
itsdangerous==2.0.1
jinja2==3.0.1
libcst==0.3.21
locust==2.2.3
markupsafe==2.0.1
msgpack==1.0.2
mypy-extensions==0.4.3
packaging==21.0
pluggy==1.0.0
promise==2.3
proto-plus==1.19.4
protobuf==3.18.1
psutil==5.8.0
py==1.10.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pydantic==1.8.2
pyparsing==2.4.7
pytest==6.2.5
pyyaml==5.4.1
pyzmq==22.3.0
requests==2.26.0
roundrobin==0.0.2
rsa==4.7.2
rx==1.6.1
six==1.16.0
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.2
typing-inspect==0.7.1
typing==3.7.4.3
urllib3==1.26.7
werkzeug==2.0.1
zope.event==4.5.0
zope.interface==5.4.0
